+++ src/main/java/com/example/kinoarena/controllers/ProjectionController.java
@@ -22,6 +22,7 @@ import com.example.kinoarena.dao.ProjectionDao;
 import com.example.kinoarena.dao.UserDao;
 import com.example.kinoarena.dto.ProjectionDto;
 import com.example.kinoarena.dto.ProjectionWithMoviesDto;
+import com.example.kinoarena.exceptions.CinemaNotFoundException;
 import com.example.kinoarena.exceptions.InvalidInputDataException;
 import com.example.kinoarena.exceptions.KinoArenaException;
 import com.example.kinoarena.exceptions.NotAdminException;
@@ -36,113 +37,95 @@ import lombok.Setter;
 @Setter
 
 @RestController
-public class ProjectionController extends BaseController{
+public class ProjectionController extends BaseController {
+
+	private static final String PROJECTION_NOT_FOUND = "There is no projection with this id!";
 
 	@Autowired
 	private ProjectionRepository projectionRepository;
-	
+
 	@Autowired
 	private ProjectionDao projectionDao;
 
-	//Works!
+	// Works!
 	// Get all projections (with the movie ids)
 	@GetMapping("/projections/movie_ids")
 	public List<Projection> getAllProjections() {
 		return projectionRepository.findAll();
 	}
 
-	//Works!
-	//Get all projections (with the movie titles)
-		@GetMapping("/projections/movie_titles")
-		public List<ProjectionWithMoviesDto> getAllProjectionsWithMovieTitles() throws KinoArenaException, SQLException {
-			return projectionDao.getAllProjectionWithMovieTitles();
-		}
-	
-	//Works!
-	//Get projection by Id
+	// Works!
+	// Get all projections (with the movie titles)
+	@GetMapping("/projections/movie_titles")
+	public List<ProjectionWithMoviesDto> getAllProjectionsWithMovieTitles() throws KinoArenaException, SQLException {
+		return projectionDao.getAllProjectionWithMovieTitles();
+	}
+
+	// Works!
+	// Get projection by Id
 	@GetMapping("/projection/{id}")
 	public Projection getProjectionById(@PathVariable Long id) throws KinoArenaException {
 		Projection projection = projectionRepository.findByProjectionId(id);
-		if(projection != null) {
+		if (projection != null) {
 			return projection;
 		} else {
 			throw new ProjectionNotFoundException("Projection with id = " + id + " not found!");
 		}
 	}
 
-	
-	// works!
-	// TODO validations
+	// Works!
 	@PostMapping("/projection/add")
-	public String addProjection(@RequestBody ProjectionDto projectionDto,HttpSession session,HttpServletRequest request ) throws KinoArenaException{
+	public void addProjection(@RequestBody ProjectionDto projectionDto, HttpSession session, HttpServletRequest request)
+			throws KinoArenaException {
 		validateLoginAdmin(request);
 		Projection projection = new Projection();
 		try {
-			    SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm");
-			    format.parse(projectionDto.getStartTime());
-			    format.parse(projectionDto.getEndTime());
+			SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm");
+			format.parse(projectionDto.getStartTime());
+			format.parse(projectionDto.getEndTime());
 			projection.setStartTime(projectionDto.getStartTime());
 			projection.setEndTime(projectionDto.getEndTime());
 			projection.setMovieId(projectionDto.getMovieId());
-			
-			if(projectionDao.checkIfProjectionExists(projectionDto.getStartTime(), projectionDto.getMovieId())) {
-				return "Projection already exists!";
+
+			if (projectionDao.checkIfProjectionExists(projectionDto.getStartTime(), projectionDto.getMovieId())) {
+				throw new KinoArenaException("Projection already exists!");
 			}
-			
-//			if(projectionRepository.findByMovieIdAndStartTime(projectionDto.getMovieId(), (projectionDto.getStartTime()))) {
-//				return "Projection already exists!";
-//			}
-//			
+
 			projectionRepository.save(projection);
-			return "Projection was added successfully!";
-			
-//		} catch(KinoArenaException e) {
-//			return e.getMessage();
-		} catch(Exception e) {
+		} catch (KinoArenaException e) {
+			throw new KinoArenaException(e.getMessage());
+		} catch (Exception e) {
 			throw new InvalidInputDataException();
 		}
 	}
-	
-	
-//	@PostMapping("/projection/find")
-//	public Projection findProjection(@RequestBody ProjectionDto projectionDto,HttpSession session,HttpServletRequest request ) throws KinoArenaException, ParseException{
-//		Projection projection = new Projection();
-////		try {
-//			    SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm");
-//			    format.parse(projectionDto.getStartTime());
-//			    format.parse(projectionDto.getEndTime());
-//			projection.setStartTime(projectionDto.getStartTime());
-//			projection.setEndTime(projectionDto.getEndTime());
-//			projection.setMovieId(projectionDto.getMovieId());
-//			return projectionRepository.findByMovieIdAndStartTime(projectionDto.getMovieId(), (projectionDto.getStartTime()));
-////		} catch(Exception e) {
-////			throw new InvalidInputDataException();
-////		}
-//	}
-	
-	
+
+	// Works!
 	@GetMapping("/projections/cinema/{id}")
-	public List<Long> getProjectionIdsByCinemaId(@PathVariable Long id) throws ProjectionNotFoundException, InvalidInputDataException, SQLException {
+	public List<Long> getProjectionIdsByCinemaId(@PathVariable Long id)
+			throws ProjectionNotFoundException, InvalidInputDataException, SQLException {
 		return projectionDao.getProjectionIdsByCinemaId(id);
 	}
-	
+
+	// Works!
 	@GetMapping("/projections/all/cinema/{id}")
 	public List<Projection> getProjectionsByCinemaId(@PathVariable Long id) throws KinoArenaException, SQLException {
 		List<Projection> projections = projectionDao.getProjectionsByCinemaId(id);
-		if(!projections.isEmpty()) {
+		if (!projections.isEmpty()) {
 			return projections;
 		} else {
 			throw new KinoArenaException("There is no projection for this cinema!");
 		}
 	}
-	
-	//DOESN'T WORK!
+
+	// DOESN'T WORK!
 	@DeleteMapping("/projections/delete/{id}")
 	public void deleteProjection(@PathVariable Long id, HttpServletRequest request) throws KinoArenaException {
 		validateLoginAdmin(request);
-		projectionRepository.deleteById(id);
+		if (projectionRepository.existsById(id)) {
+			projectionRepository.deleteById(id);
+		} else {
+			throw new ProjectionNotFoundException(PROJECTION_NOT_FOUND);
+		}
 	}
-	
-	
 
 }
